separated temporary view frames
